--- backend/src/main/java/com/example/rabbithell/auth/adapter/in/web/AuthController.java
@@ -1,50 +1,0 @@
-package com.example.rabbithell.auth.adapter.in.web;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.example.rabbithell.auth.application.AuthService;
-import com.example.rabbithell.auth.domain.AuthUser;
-import com.example.rabbithell.auth.dto.LoginRequest;
-import com.example.rabbithell.auth.dto.LoginResponse;
-import com.example.rabbithell.auth.dto.SignupRequest;
-import com.example.rabbithell.auth.dto.TokenResponse;
-
-import lombok.RequiredArgsConstructor;
-
-@RestController
-@RequestMapping("/auth")
-@RequiredArgsConstructor
-public class AuthController {
-
-    private final AuthService authService;
-
-    @PostMapping("/signup")
-    public ResponseEntity<Void> signup(@RequestBody SignupRequest request) {
-        authService.signup(request);
-        return ResponseEntity.ok().build();
-    }
-
-    @PostMapping("/login")
-    public ResponseEntity<LoginResponse> login(@RequestBody LoginRequest request) {
-        return ResponseEntity.ok(authService.login(request.email(), request.password()));
-    }
-
-    @PostMapping("/reissue")
-    public ResponseEntity<TokenResponse> reissue(@RequestHeader("Authorization") String bearerToken) {
-        String refreshToken = bearerToken.replace("Bearer ", "");
-        return ResponseEntity.ok(authService.reissue(refreshToken));
-    }
-
-    @PostMapping("/logout")
-    public ResponseEntity<Void> logout(@AuthenticationPrincipal AuthUser authUser) {
-        authService.logout(authUser.getUserId());
-        return ResponseEntity.ok().build();
-    }
-
-}
