--- backend/src/main/java/com/example/rabbithell/auth/dto/LoginRequest.java
@@ -1,6 +1,0 @@
-package com.example.rabbithell.auth.dto;
-
-public record LoginRequest(
-    String email,
-    String password
-) {}
