--- backend/src/main/java/com/example/rabbithell/auth/dto/LoginResponse.java
@@ -1,6 +1,0 @@
-package com.example.rabbithell.auth.dto;
-
-public record LoginResponse(
-    String accessToken,
-    String refreshToken
-) {}
