--- backend/src/main/java/com/example/rabbithell/auth/dto/SignupRequest.java
@@ -1,7 +1,0 @@
-package com.example.rabbithell.auth.dto;
-
-public record SignupRequest(
-    String email,
-    String password,
-    String role
-) {}
