--- backend/src/main/java/com/example/rabbithell/auth/dto/TokenResponse.java
@@ -1,6 +1,0 @@
-package com.example.rabbithell.auth.dto;
-
-public record TokenResponse(
-    String accessToken,
-    String refreshToken
-) {}
