--- backend/src/main/java/com/example/rabbithell/auth/infrastructure/jwt/JwtTokenProvider.java
@@ -1,88 +1,0 @@
-package com.example.rabbithell.auth.infrastructure.jwt;
-
-import java.nio.charset.StandardCharsets;
-import java.security.Key;
-import java.util.Date;
-
-import org.springframework.stereotype.Component;
-
-import com.example.rabbithell.auth.port.out.JwtTokenPort;
-
-import io.jsonwebtoken.JwtException;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-import io.jsonwebtoken.security.Keys;
-import jakarta.annotation.PostConstruct;
-import lombok.RequiredArgsConstructor;
-
-@Component
-@RequiredArgsConstructor
-public class JwtTokenProvider implements JwtTokenPort {
-
-    private final JwtProperties properties;
-    private Key key;
-
-    @PostConstruct
-    public void init() {
-        this.key = Keys.hmacShaKeyFor(properties.getSecret().getBytes(StandardCharsets.UTF_8));
-    }
-
-    @Override
-    public String generateAccessToken(String userId, String role) {
-        return Jwts.builder()
-            .setSubject(userId)
-            .claim("role", role)
-            .setIssuedAt(new Date())
-            .setExpiration(new Date(System.currentTimeMillis() + getAccessTokenExpireMillis()))
-            .signWith(key, SignatureAlgorithm.HS256)
-            .compact();
-    }
-
-    @Override
-    public String generateRefreshToken(String userId) {
-        return Jwts.builder()
-            .setSubject(userId)
-            .setIssuedAt(new Date())
-            .setExpiration(new Date(System.currentTimeMillis() + getRefreshTokenExpireMillis()))
-            .signWith(key, SignatureAlgorithm.HS256)
-            .compact();
-    }
-
-    @Override
-    public boolean validateToken(String token) {
-        try {
-            Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token);
-            return true;
-        } catch (JwtException | IllegalArgumentException e) {
-            return false;
-        }
-    }
-
-    @Override
-    public String extractSubject(String token) {
-        return Jwts.parserBuilder()
-            .setSigningKey(key)
-            .build()
-            .parseClaimsJws(token)
-            .getBody()
-            .getSubject();
-    }
-
-    @Override
-    public String extractRole(String token) {
-        return Jwts.parserBuilder()
-            .setSigningKey(key)
-            .build()
-            .parseClaimsJws(token)
-            .getBody()
-            .get("role", String.class);
-    }
-
-    private long getAccessTokenExpireMillis() {
-        return 1000 * 60 * properties.getToken().getAccess().getMinute();
-    }
-
-    private long getRefreshTokenExpireMillis() {
-        return 1000 * 60 * properties.getToken().getRefresh().getMinute();
-    }
-}
