+++ backend/src/main/java/com/example/rabbithell/common/exception/BaseException.java
@@ -1,0 +1,11 @@
+package com.example.rabbithell.common.exception;
+
+import org.springframework.http.HttpStatus;
+
+public abstract class BaseException extends RuntimeException {
+    public abstract Enum<?> getErrorCode();
+
+    public abstract HttpStatus getHttpStatus();
+
+    public abstract String getMessage();
+}
