+++ backend/src/main/java/com/example/rabbithell/common/exception/dto/ValidationError.java
@@ -1,0 +1,12 @@
+package com.example.rabbithell.common.exception.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class ValidationError {
+    private String field;
+    private String message;
+    private String code;
+}
