+++ backend/src/main/java/com/example/rabbithell/config/RedisConfig.java
@@ -1,0 +1,27 @@
+package com.example.rabbithell.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.redis.connection.RedisConnectionFactory;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.serializer.StringRedisSerializer;
+
+@Configuration
+public class RedisConfig {
+
+    @Bean
+    public RedisTemplate<String, String> redisTemplate(RedisConnectionFactory factory) {
+
+        RedisTemplate<String, String> template = new RedisTemplate<>();
+
+        template.setConnectionFactory(factory);
+
+        template.setKeySerializer(new StringRedisSerializer());
+
+        template.setValueSerializer(new StringRedisSerializer());
+
+        return template;
+
+    }
+
+}
