+++ backend/src/main/java/com/example/rabbithell/domain/auth/domain/AuthUser.java
@@ -1,0 +1,56 @@
+package com.example.rabbithell.domain.auth.domain;
+
+import java.util.Collection;
+import java.util.Collections;
+
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import lombok.Getter;
+
+@Getter
+public class AuthUser implements UserDetails {
+
+    private final String userId;
+    private final String role;
+
+    public AuthUser(String userId, String role) {
+        this.userId = userId;
+        this.role = role;
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return Collections.emptyList(); // 필요 시 권한 리스트 반환
+    }
+
+    @Override
+    public String getPassword() {
+        return null;
+    }
+
+    @Override
+    public String getUsername() {
+        return userId;
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
+}
