+++ backend/src/main/java/com/example/rabbithell/domain/auth/dto/request/LoginRequest.java
@@ -1,0 +1,6 @@
+package com.example.rabbithell.domain.auth.dto.request;
+
+public record LoginRequest(
+    String email,
+    String password
+) {}
