+++ backend/src/main/java/com/example/rabbithell/domain/auth/dto/request/SignupRequest.java
@@ -1,0 +1,7 @@
+package com.example.rabbithell.domain.auth.dto.request;
+
+public record SignupRequest(
+    String email,
+    String password,
+    String role
+) {}
