+++ backend/src/main/java/com/example/rabbithell/domain/auth/dto/response/LoginResponse.java
@@ -1,0 +1,6 @@
+package com.example.rabbithell.domain.auth.dto.response;
+
+public record LoginResponse(
+    String accessToken,
+    String refreshToken
+) {}
