+++ backend/src/main/java/com/example/rabbithell/domain/auth/dto/response/TokenResponse.java
@@ -1,0 +1,6 @@
+package com.example.rabbithell.domain.auth.dto.response;
+
+public record TokenResponse(
+    String accessToken,
+    String refreshToken
+) {}
