+++ backend/src/main/java/com/example/rabbithell/infrastructure/security/jwt/JwtProperties.java
@@ -1,0 +1,37 @@
+package com.example.rabbithell.infrastructure.security.jwt;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Component
+@ConfigurationProperties(prefix = "spring.jwt")
+@Getter
+@Setter
+public class JwtProperties {
+    private String secret;
+    private Token token;
+
+    @Getter
+    @Setter
+    public static class Token {
+        private String header;
+        private String prefix;
+        private Access access;
+        private Refresh refresh;
+
+        @Getter
+        @Setter
+        public static class Access {
+            private long minute;
+        }
+
+        @Getter
+        @Setter
+        public static class Refresh {
+            private long minute;
+        }
+    }
+}
