+++ backend/src/main/resources/application-local.yml
@@ -23,7 +23,7 @@ spring:
     # JPA 설정
     jpa:
         hibernate:
-            ddl-auto: update
+            ddl-auto: create
         open-in-view: false
         properties:
             hibernate:
